# docs/openapi.product-surface.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "VOS3000 Portal Product API Surface",
  "version": "1.0.0",
  "description": "Generated from the supplied Admin and Client product specifications. VOS-specific writes remain capability-gated until verified."
 },
 "servers": [
  {
   "url": "http://localhost:5026"
  }
 ],
 "paths": {
  "/api/v1/admin/auth/login": {
   "post": {
    "summary": "POST /api/v1/admin/auth/login",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/auth/mfa/verify": {
   "post": {
    "summary": "POST /api/v1/admin/auth/mfa/verify",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/auth/password/request": {
   "post": {
    "summary": "POST /api/v1/admin/auth/password/request",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/auth/password/reset": {
   "post": {
    "summary": "POST /api/v1/admin/auth/password/reset",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/me/sessions": {
   "get": {
    "summary": "GET /api/v1/admin/me/sessions",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
  "/api/v1/admin/me/sessions/{id}": {
   "delete": {
    "summary": "DELETE /api/v1/admin/me/sessions/{id}",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
...
  "/api/v1/admin/noc/stream": {
   "get": {
    "summary": "GET /api/v1/admin/noc/stream",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
...
  "/api/v1/admin/customers/{id}/adjustments": {
   "post": {
    "summary": "POST /api/v1/admin/customers/{id}/adjustments",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/diagnostics/call-analysis/{serial}": {
   "get": {
    "summary": "GET /api/v1/admin/diagnostics/call-analysis/{serial}",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "parameters": [
     {
      "name": "serial",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
  "/api/v1/admin/diagnostics/registration-analysis": {
   "get": {
    "summary": "GET /api/v1/admin/diagnostics/registration-analysis",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
  "/api/v1/admin/cdr/recent": {
   "get": {
    "summary": "GET /api/v1/admin/cdr/recent",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
  "/api/v1/admin/cdr": {
   "get": {
    "summary": "GET /api/v1/admin/cdr",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
  "/api/v1/admin/cdr/exports": {
   "post": {
    "summary": "POST /api/v1/admin/cdr/exports",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/cdr/{id}": {
   "get": {
    "summary": "GET /api/v1/admin/cdr/{id}",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
...
  "/api/v1/admin/rates/groups": {
   "get": {
    "summary": "GET /api/v1/admin/rates/groups",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   },
   "post": {
    "summary": "POST /api/v1/admin/rates/groups",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/rates/groups/{id}/rates": {
   "get": {
    "summary": "GET /api/v1/admin/rates/groups/{id}/rates",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   },
   "put": {
    "summary": "PUT /api/v1/admin/rates/groups/{id}/rates",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/rates/lookup": {
   "get": {
    "summary": "GET /api/v1/admin/rates/lookup",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
...
  "/api/v1/auth/login": {
   "post": {
    "summary": "POST /api/v1/auth/login",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
  "/api/v1/me/mfa": {
   "post": {
    "summary": "POST /api/v1/me/mfa",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
  "/api/v1/me/sessions": {
   "get": {
    "summary": "GET /api/v1/me/sessions",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
  "/api/v1/me/sessions/{id}": {
   "delete": {
    "summary": "DELETE /api/v1/me/sessions/{id}",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
...
  "/api/v1/deposits": {
   "post": {
    "summary": "POST /api/v1/deposits",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   }
  },
  "/api/v1/deposits/{id}": {
   "get": {
    "summary": "GET /api/v1/deposits/{id}",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
...
  "/api/v1/cdr": {
   "get": {
    "summary": "GET /api/v1/cdr",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
  "/api/v1/cdr/exports": {
   "post": {
    "summary": "POST /api/v1/cdr/exports",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "requestBody": {
     "required": false,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true
       }
      }
     }
    }
   },
   "get": {
    "summary": "GET /api/v1/cdr/exports",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
  "/api/v1/cdr/{id}": {
   "get": {
    "summary": "GET /api/v1/cdr/{id}",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
  "/api/v1/cdr/recent": {
   "get": {
    "summary": "GET /api/v1/cdr/recent",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
...
  "/api/v1/rates": {
   "get": {
    "summary": "GET /api/v1/rates",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
  "/api/v1/rates/lookup": {
   "get": {
    "summary": "GET /api/v1/rates/lookup",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   }
  },
...
  "/api/v1/admin/settings/support": {
   "get": {
    "summary": "GET /api/v1/admin/settings/support",
    "description": "Global support contacts (Telegram + Teams) for the client Support FAB. Admin only.",
    "responses": {
     "200": {
      "description": "Successful portal response"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden"
     }
    }
   },
   "put": {
    "summary": "PUT /api/v1/admin/settings/support",
    "description": "Validate, persist and audit global support contacts; invalidates the support:config cache.",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "required": [
         "enabled",
         "telegram",
         "teams"
        ],
        "properties": {
         "enabled": {
          "type": "boolean"
         },
         "label": {
          "type": "string",
          "maxLength": 120
         },
         "telegram": {
          "type": "object",
          "properties": {
           "enabled": {
            "type": "boolean"
           },
           "handle": {
            "type": "string",
            "example": "@my_support_bot"
           }
          }
         },
         "teams": {
          "type": "object",
          "properties": {
           "enabled": {
            "type": "boolean"
           },
           "id": {
            "type": "string",
            "example": "support@example.com"
           }
          }
         }
        },
        "example": {
         "enabled": true,
         "label": "Need help?",
         "telegram": {
          "enabled": true,
          "handle": "@my_support_bot"
         },
         "teams": {
          "enabled": true,
          "id": "support@example.com"
         }
        }
       }
      }
     },
     "responses": null
    },
    "responses": {
     "200": {
      "description": "Saved and audited"
     },
     "400": {
      "description": "VALIDATION_ERROR with details.fields[]"
     },
     "401": {
      "description": "Authentication required"
     },
     "403": {
      "description": "Forbidden / invalid origin"
     }
    }
   }
  },
...
 }
}